Added Ze Pipeline For The Whole Thingy
Now we're rollin'

Basically, NYAAAAAAA

Wait, no, no, focus! Hocus pocus don't lose focuss!

. . .

Anyway, I made a MainPipeline class, what it does is that it checks every 3 minutes for update in the database. When an update happens, it'll do some AI powered backflips and populate it.

That's it

. . .

This might work, who knows?

Let's gooooooo~